+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/Action.java
@@ -15,17 +15,22 @@
  */
 package com.google.android.exoplayer2.testutil;
 
-import android.os.Handler;
 import android.util.Log;
 import android.view.Surface;
+import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
+import com.google.android.exoplayer2.PlayerMessage;
+import com.google.android.exoplayer2.PlayerMessage.Target;
 import com.google.android.exoplayer2.SimpleExoPlayer;
 import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.testutil.ActionSchedule.ActionNode;
+import com.google.android.exoplayer2.testutil.ActionSchedule.PlayerRunnable;
+import com.google.android.exoplayer2.testutil.ActionSchedule.PlayerTarget;
 import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
+import com.google.android.exoplayer2.util.HandlerWrapper;
 
 /**
  * Base class for actions to perform during playback tests.
@@ -53,15 +58,19 @@ public abstract class Action {
    * @param handler The handler to use to pass to the next action.
    * @param nextAction The next action to schedule immediately after this action finished.
    */
-  public final void doActionAndScheduleNext(SimpleExoPlayer player,
-      MappingTrackSelector trackSelector, Surface surface, Handler handler, ActionNode nextAction) {
+  public final void doActionAndScheduleNext(
+      SimpleExoPlayer player,
+      MappingTrackSelector trackSelector,
+      Surface surface,
+      HandlerWrapper handler,
+      ActionNode nextAction) {
     Log.i(tag, description);
     doActionAndScheduleNextImpl(player, trackSelector, surface, handler, nextAction);
   }
 
   /**
    * Called by {@link #doActionAndScheduleNext(SimpleExoPlayer, MappingTrackSelector, Surface,
-   * Handler, ActionNode)} to perform the action and to schedule the next action node.
+   * HandlerWrapper, ActionNode)} to perform the action and to schedule the next action node.
    *
    * @param player The player to which the action should be applied.
    * @param trackSelector The track selector to which the action should be applied.
@@ -69,8 +78,12 @@ public abstract class Action {
    * @param handler The handler to use to pass to the next action.
    * @param nextAction The next action to schedule immediately after this action finished.
    */
-  protected void doActionAndScheduleNextImpl(SimpleExoPlayer player,
-      MappingTrackSelector trackSelector, Surface surface, Handler handler, ActionNode nextAction) {
+  protected void doActionAndScheduleNextImpl(
+      SimpleExoPlayer player,
+      MappingTrackSelector trackSelector,
+      Surface surface,
+      HandlerWrapper handler,
+      ActionNode nextAction) {
     doActionImpl(player, trackSelector, surface);
     if (nextAction != null) {
       nextAction.schedule(player, trackSelector, surface, handler);
@@ -79,14 +92,14 @@ public abstract class Action {
 
   /**
    * Called by {@link #doActionAndScheduleNextImpl(SimpleExoPlayer, MappingTrackSelector, Surface,
-   * Handler, ActionNode)} to perform the action.
+   * HandlerWrapper, ActionNode)} to perform the action.
    *
    * @param player The player to which the action should be applied.
    * @param trackSelector The track selector to which the action should be applied.
    * @param surface The surface to use when applying actions.
    */
-  protected abstract void doActionImpl(SimpleExoPlayer player, MappingTrackSelector trackSelector,
-      Surface surface);
+  protected abstract void doActionImpl(
+      SimpleExoPlayer player, MappingTrackSelector trackSelector, Surface surface);
 
   /**
    * Calls {@link Player#seekTo(long)} or {@link Player#seekTo(int, long)}.
@@ -345,7 +358,63 @@ public abstract class Action {
         Surface surface) {
       player.setShuffleModeEnabled(shuffleModeEnabled);
     }
+  }
 
+  /** Calls {@link ExoPlayer#createMessage(Target)} and {@link PlayerMessage#send()}. */
+  public static final class SendMessages extends Action {
+
+    private final Target target;
+    private final int windowIndex;
+    private final long positionMs;
+    private final boolean deleteAfterDelivery;
+
+    /**
+     * @param tag A tag to use for logging.
+     * @param target A message target.
+     * @param positionMs The position at which the message should be sent, in milliseconds.
+     */
+    public SendMessages(String tag, Target target, long positionMs) {
+      this(
+          tag,
+          target,
+          /* windowIndex= */ C.INDEX_UNSET,
+          positionMs,
+          /* deleteAfterDelivery= */ true);
+    }
+
+    /**
+     * @param tag A tag to use for logging.
+     * @param target A message target.
+     * @param windowIndex The window index at which the message should be sent, or {@link
+     *     C#INDEX_UNSET} for the current window.
+     * @param positionMs The position at which the message should be sent, in milliseconds.
+     * @param deleteAfterDelivery Whether the message will be deleted after delivery.
+     */
+    public SendMessages(
+        String tag, Target target, int windowIndex, long positionMs, boolean deleteAfterDelivery) {
+      super(tag, "SendMessages");
+      this.target = target;
+      this.windowIndex = windowIndex;
+      this.positionMs = positionMs;
+      this.deleteAfterDelivery = deleteAfterDelivery;
+    }
+
+    @Override
+    protected void doActionImpl(
+        final SimpleExoPlayer player, MappingTrackSelector trackSelector, Surface surface) {
+      if (target instanceof PlayerTarget) {
+        ((PlayerTarget) target).setPlayer(player);
+      }
+      PlayerMessage message = player.createMessage(target);
+      if (windowIndex != C.INDEX_UNSET) {
+        message.setPosition(windowIndex, positionMs);
+      } else {
+        message.setPosition(positionMs);
+      }
+      message.setHandler(new android.os.Handler());
+      message.setDeleteAfterDelivery(deleteAfterDelivery);
+      message.send();
+    }
   }
 
   /**
@@ -388,8 +457,11 @@ public abstract class Action {
     }
 
     @Override
-    protected void doActionAndScheduleNextImpl(final SimpleExoPlayer player,
-        final MappingTrackSelector trackSelector, final Surface surface, final Handler handler,
+    protected void doActionAndScheduleNextImpl(
+        final SimpleExoPlayer player,
+        final MappingTrackSelector trackSelector,
+        final Surface surface,
+        final HandlerWrapper handler,
         final ActionNode nextAction) {
       if (nextAction == null) {
         return;
@@ -432,8 +504,11 @@ public abstract class Action {
     }
 
     @Override
-    protected void doActionAndScheduleNextImpl(final SimpleExoPlayer player,
-        final MappingTrackSelector trackSelector, final Surface surface, final Handler handler,
+    protected void doActionAndScheduleNextImpl(
+        final SimpleExoPlayer player,
+        final MappingTrackSelector trackSelector,
+        final Surface surface,
+        final HandlerWrapper handler,
         final ActionNode nextAction) {
       if (nextAction == null) {
         return;
@@ -472,8 +547,11 @@ public abstract class Action {
     }
 
     @Override
-    protected void doActionAndScheduleNextImpl(final SimpleExoPlayer player,
-        final MappingTrackSelector trackSelector, final Surface surface, final Handler handler,
+    protected void doActionAndScheduleNextImpl(
+        final SimpleExoPlayer player,
+        final MappingTrackSelector trackSelector,
+        final Surface surface,
+        final HandlerWrapper handler,
         final ActionNode nextAction) {
       if (nextAction == null) {
         return;
@@ -514,8 +592,11 @@ public abstract class Action {
     }
 
     @Override
-    protected void doActionAndScheduleNextImpl(final SimpleExoPlayer player,
-        final MappingTrackSelector trackSelector, final Surface surface, final Handler handler,
+    protected void doActionAndScheduleNextImpl(
+        final SimpleExoPlayer player,
+        final MappingTrackSelector trackSelector,
+        final Surface surface,
+        final HandlerWrapper handler,
         final ActionNode nextAction) {
       if (nextAction == null) {
         return;
@@ -555,6 +636,9 @@ public abstract class Action {
     @Override
     protected void doActionImpl(SimpleExoPlayer player, MappingTrackSelector trackSelector,
         Surface surface) {
+      if (runnable instanceof PlayerRunnable) {
+        ((PlayerRunnable) runnable).setPlayer(player);
+      }
       runnable.run();
     }